getting closer.. sync is working [ish]

+++ include/view/waveformview.hpp
@@ -38,7 +38,7 @@ namespace DataJockey {
             QGraphicsLineItem * mCursor;
             Audio::BeatBuffer mBeatBuffer;
             QList<QGraphicsLineItem * > mBeatLines;
-            unsigned int mSampleRate;
+            double mSampleRate;
       };
    }
 }
